% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_main.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Estimate confidence intervals for the coefficients of logistic regression}
\usage{
bootstrap_ci(y, X, alpha = 0.05, rounds = 20)
}
\arguments{
\item{y}{A \code{double} value of the vector containing the response of interest.}

\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{alpha}{A \code{double} value indicating the significance level of the confidence intervals.}

\item{rounds}{A \code{double} value indicating how many times to perform the bootstrap.}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{boot_ci_int}{The estimated confidence interval for the coefficient of the intercept}
 \item{boot_ci_slope}{The estimated confidence interval for the coefficient of the slope}
}
}
\description{
This function bootstraps the input data to create sample populations to estimate the coefficients of logistic regression.
}
\author{
Micheal Stewart Jackson
}
