% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_main.R
\name{log_curve}
\alias{log_curve}
\title{Create a fitted logistic curve}
\usage{
log_curve(X, y, beta)
}
\arguments{
\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{y}{A \code{double} value of the vector containing the response of interest.}

\item{beta}{A \code{double} value that has previously been found with the ls_optim function.}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{model}{fit of the logistic regression model}
}
}
\description{
This function shows a visual of the logistic regression
}
\author{
Kayla Gallman
}
