% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_main.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Create a plot of accuracy from a confusion matrix}
\usage{
plot_metrics(y, X, alpha = 0.5)
}
\arguments{
\item{y}{A \code{double} value of the vector containing the response of interest.}

\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{alpha}{A \code{double} value indicating the significance level of the confidence intervals for the logistic regression coefficients.}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{model}{}
}
}
\description{
This function shows plot of accuracy from a confusion matrix evaluated over a grid of cut-off values for prediction going from 0.1 to 0.9
}
\author{
Kayla Gallman
}
